+++ changelog.md
@@ -1,4 +1,5 @@
 [![Downloads](https://img.shields.io/github/downloads/probablyraging/steam-game-idler/1.5.25/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/probablyraging/steam-game-idler/releases/download/1.5.25/Steam.Game.Idler_1.5.25_x64_en-US.msi)
 
 ## Changelog
-- Always clear stored Steam web API key when logging out of SGI
+- Added `Free game notifications` to `settings > general`
+  - Receive native Windows notifications when free Steam games are available to add to your account